Add convenience functions for face data.
- Allows obtaining the value currently being sent on a face. This
  simplifies and reduces the memory usage of any program that needs to
  modify the value being sent based on the existing one (for instance,
  setting bits. Without this, one needs to keep track of the values
  themselves.
- Allows access to data for datagrams pending to be sent on a given
  face. Juts as above, this simplifies and reduces (here by a lot)
  memory usage of any programs that need access to the outgoing datagram
  (say, to use the data for send retries) and even to the fact that
  there is a pending datagram (no need to keep track of this to prevent
  some code from overwritting the data there unwittingly in a single
  loop iteration.

All in all, this simplifies some usage scenarios for certain programs,
add a nice simimetry with the existing methods and, above all, cam
pottentially save a considerable amount of memory usage form some types
of programs (up to 102 bytes can be saved.

Adding the fact that if these functions are not used they take no extra
storage space, this looks like a no-brainer to me.